[GStreamer][MSE] Fix 'comparison between signed and unsigned integer expressions' warning
Reviewed by Žan Doberšek.

* platform/graphics/gstreamer/mse/PlaybackPipeline.cpp:
(WebCore::PlaybackPipeline::flush):

Canonical link: https://commits.webkit.org/186393@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@213636 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/graphics/gstreamer/mse/PlaybackPipeline.cpp
@@ -425,7 +425,7 @@ void PlaybackPipeline::flush(AtomicString trackId)
 
     GST_TRACE("Position: %" GST_TIME_FORMAT, GST_TIME_ARGS(position));
 
-    if (position == GST_CLOCK_TIME_NONE) {
+    if (static_cast<guint64>(position) == GST_CLOCK_TIME_NONE) {
         GST_TRACE("Can't determine position, avoiding flush");
         return;
     }
